+++ dangerousgary.mjs
@@ -59,7 +59,7 @@ Hooks.once('init', async function () {
 })
 
 Hooks.on("hotbarDrop", (bar, data, slot) => {
-  if (["rollDamage", "Item"].includes(data.type)) {
+  if (["rollDamage", "rollSave", "Item"].includes(data.type)) {
     Macros.createDangerousGaryMacro(data, slot)
     return false
   }